edited routes id tests

+++ tests/api/services/user.service.js
@@ -2,6 +2,6 @@ import BaseService from "../lib/base.service.js";
 
 export default class UserService extends BaseService {
     deleteById(id) {
-        return this.client.delete(`/users/${id}`)
+        return this.client.delete(`/api/auth/users/${id}`)
     }
 }
